+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - PR #2651 Remove nvidia driver installation from ci/cpu/build.sh
 - PR #2697 Ensure csv reader sets datetime column time units
 - PR #2698 Return RangeIndex from contiguous slice of RangeIndex
+- PR #2672 Fix null and integer handling in round
 
 
 # cuDF 0.9.0 (Date TBD)